Created urls util file and migrated relevant code there
